chore: update training pipeline, models, and stability fixes (removed large checkpoints)

+++ cmd/inspect/inspect_model.go
@@ -14,7 +14,7 @@ import (
 type CheckpointHeader struct {
 	StepCount       int
 	Version         string
-	Commitment      float64
+	Commitment      float32
 	TokensProcessed int64
 	TotalDuration   string
 	LastProfile     nn.TrainingProfile
